KAFKA-1042 logSegments(from, to) Misses first segment.
I think this patch also fixes the continual timing problems we have had in log tests. The root cause
was that we weren't passing through the clock instance so we were mixing instances of MockTime and SystemTime.
This worked only because MockTime initializes to SystemTime.milliseconds so as long as the test took less than 1 ms it worked!