Remove non-functional variable definition in log4j.properties
In log4j.properties, a property kafka.logs.dir was defined. However, modifying this property has no effect because log4j.properties does not support variable substitution in this manner. Rather ${kafka.logs.dir} is looked up as a System property, which is set in bin/kafka-run-class.sh, and must currently be changed there if the log4j logging location is to be modified.