KAFKA-1595; Remove deprecated and slower scala JSON parser from kafka.consumer.TopicCount
A thin wrapper over Jackson's Tree Model API is used as the replacement. This wrapper
increases safety while providing a simple, but powerful API through the usage of the
`DecodeJson` type class. Even though this has a maintenance cost, it makes the API
much more convenient from Scala. A number of tests were added to verify the
behaviour of this wrapper.

The Scala module for Jackson doesn't provide any help for our current usage, so we don't
depend on it.

An attempt has been made to maintain the existing behaviour regarding when exceptions
are thrown. There are a number of cases where `JsonMappingException` will be thrown
instead of `ClassCastException`, however. It is expected that users would not try to catch
`ClassCastException`.