KAFKA-1595; Remove deprecated and slower scala JSON parser from kafka.consumer.TopicCount
A combination of spray-json's AST combined with jawn's parser are used as the replacement.
Note that both libraries have no dependencies and are relatively simple. We use `jawn` for
its performance, but it could be dropped by changing one line in `Json.parseFull`.

An attempt has been made to maintain the existing behaviour regarding when exceptions
are thrown. There are a number of cases where `DeserializationException` will be thrown
instead of `ClassCastException`, however. It is expected that users would not try to catch
`ClassCastException`.