Restrict offset format.
Restrict offsets to be serialized via schemaless objects, but validate that they
are maps with string keys and primitive values. This is enforced at the Copycat
level rather than validating a Schema.

Additionally, parameterize copycat system test to work with and without schemas,
and to validate both JSON and String converters.