Added a new ProcessorTopologyTestDriver class that makes it easier to unit test a ProcessorTopology.
The new test driver class uses a supplied TopologyBuilder and a StreamingConfig to set up a
topology using the existing StreamTask, which is the same component used in the Kafka Streams
framework for each thread. However, rather than rely upon a real Kafka broker, the test driver
simulates a broker with mock consumers and producers. This required a few changes to
StreamTask and StreamThread so that the latter can pass in the actual consumer used to restore
the state; this allows the test driver to pass a mock consumer into StreamTask. Without this
change, the test driver could not run topologies that used state stores.