KAFKA-2548; kafka-merge-pr tool fails to update JIRA with fix version 0.9.0.0
Simplified the logic to choose the default fix version. We just hardcode
it for `trunk` and try to compute it based on the branch name for the
rest.

Removed logic that tries to handle forked release branches as it
seems to be specific to how the Spark project handles their JIRA.

Author: Ismael Juma <ismael@juma.me.uk>

Reviewers: Guozhang Wang

Closes #238 from ijuma/kafka-2548-merge-pr-tool-4-segment-fix-version