MINOR: Expose ReplicaManager gauges
There are several gauges in core that are registered but cannot be accessed programmatically. For example, gauges "LeaderCount", "PartitionCount", "UnderReplicatedParittions" are all registered in ReplicaManager.scala but there is no way to access them programmatically if one has access to the kafka.server object. Other metrics,  such as isrExpandRate (also in ReplicaManager.scala) can be accessed. The solution here is trivial, add a var <variable name> in front of newGauge, as shown below
val partitionCount newGauge(
     "PartitionCount",
     new Gauge[Int] {
       def value = allPartitions.size
     }
)

Author: Eno Thereska <eno.thereska@gmail.com>

Reviewers: Ismael Juma, Guozhang Wang

Closes #364 from enothereska/gauges