KAFKA-2937; Disable the leaderIsr check if the topic is to be deleted.
The check was implemented in KAFKA-340 : If we are shutting down a broker when the ISR of a partition includes only that broker, we could lose some messages that have been previously committed. For clean shutdown, we need to guarantee that there is at least 1 other broker in ISR after the broker is shut down.

When we are deleting the topic, this check can be avoided.

Author: Mayuresh Gharat <mgharat@mgharat-ld1.linkedin.biz>

Reviewers: Dong Lin <lindong28@gmail.com>, Rajini Sivaram <rajinisivaram@googlemail.com>, Jun Rao <junrao@gmail.com>

Closes #729 from MayureshGharat/kafka-2937

(cherry picked from commit a788c65f02ae493abc27fe63c6d181e6a9b8edfe)
Signed-off-by: Jun Rao <junrao@gmail.com>