MINOR: Fix bug in `waitUntilLeaderIsElectedOrChanged` and simplify result type
Also disable a couple of tests that were passing incorrectly until KAFKA-3096 is fixed.

The bug was for the following case:

`leader.isDefined && oldLeaderOpt.isEmpty && newLeaderOpt.isDefined && newLeaderOpt.get != leader.get`

We would consider it a successful election even though we should not.

I also changed the result type as we never return `None` (we throw an exception instead).