KAFKA-3152; kafka-acl doesn't allow space in principal name (backport to 0.9.0)
This is a backport of the trunk PR and it excludes the test
changes due to conflicts and the fact that the changes
were not directly related to the bug in the end
(it was something we did not test for, but the non-shell code
was already correct).

Details:

* Add quotes to `$` in shell scripts
This is necessary for correct processing of quotes in the
user command.

* Minor improvements to AclCommand messages

Author: Ismael Juma <ismael@juma.me.uk>

Reviewers: Jun Rao <junrao@gmail.com>

Closes #820 from ijuma/kafka-3152-backport-kafka-acl-space-in-principal-name