HOTFIX: Include RocksDB dependency in release tarballs
Without this change `./gradlew releaseTarGz` (and its variants) will not include the RocksDB jar, which is required for Kafka Streams, in Kafka's `libs/` folder.  The impact is that any Streams job will fail when it runs against a broker that was installed via a release tarball.

guozhangwang junrao : please review.

Author: Michael G. Noll <michael@confluent.io>

Reviewers: Jun Rao <junrao@gmail.com>

Closes #1007 from miguno/trunk-rocksdb-fixes