KAFKA-3619: File handles are leaked on .lock files of ProcessorStateManager
Kafka Streams seems to hold file handles on the `.lock` files for the state dirs, resulting in an explosion of filehandles over time. Running `lsof` shows the number of open filehandles on the `.lock` file increasing rapidly over time. In a separate test project, I reproduced the issue and determined that in order for the filehandle to be relinquished the `FileChannel` instance must be properly closed. Applying this patch seems to resolve the issue in my job.

Author: Greg Fodor <gfodor@gmail.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #1267 from gfodor/bug/state-lock-filehandle-leak