Fix NPE in ConsumerGroupCommand and ConsumerOffsetChecker
The host and port entries under /brokers/ids/<bid> gets filled only for PLAINTEXT security protocol. For other protocols the host is null
and the actual endpoint is under "endpoints". This causes NPE when running the consumer group and offset checker scripts in a
kerberized env. By always reading the host and port values from the "endpoint", a more meaningful exception would be thrown rather than a NPE.