fix NPE in org.apache.kafka.clients.consumer.RoundRobinAssignor when topic metadata not found
AbstractPartitionAssignor.assign has an ambiguous line in its documentation:
@param partitionsPerTopic The number of partitions for each subscribed topic (may be empty for some topics)

Does empty mean the topic has an entry with value zero, or that the entry is excluded from the map altogether? The current implementation in AbstractPartitionAssignor excludes the entry from partitionsPerTopic if the topic isn't in the metadata.

RoundRobinAssignorTest.testOneConsumerNonexistentTopic interprets emptiness as providing the topic with a zero value.
RangeAssignor interprets emptiness as excluding the entry from the map.
RangeAssignorTest.testOneConsumerNonexistentTopic interprets emptiness as providing the topic with a zero value.

This implementation chooses to solve the NPE by deciding to exclude topics from partitionsPerTopic when the topic is not in the metadata.