KAFKA-3587: LogCleaner fails due to incorrect offset map computation on a replica
Removed the over pessimistic require and instead attempt to fill the dedup buffer. Use the (only) map until full;
this may allow to process all dirty segment (optimism) or may happen in the middle of a dirt segment.
In either case, do compaction using the map loaded that way.

This patch was developed with @edoardocomar