HOTFIX: KAFKA-3160 follow-up, catch decompression errors in constructor
After testing KAFKA-3160 a bit more, I found that the error code was not being set properly in ProduceResponse. This happened because the validation error is raised in the CompressionFactory constructor, which was not wrapped in a try / catch.

ijuma junrao

(This contribution is my original work and I license the work under Apache 2.0.)

Author: Dana Powers <dana.powers@gmail.com>
Author: Ismael Juma <ismael@juma.me.uk>

Reviewers: Jun Rao <junrao@gmail.com>, Gwen Shapira <cshapi@gmail.com>, Ismael Juma <ismael@juma.me.uk>

Closes #1344 from dpkp/decompress_error_code

(cherry picked from commit 4331bf4ff2d0e7ab1a24ea29382897162c1ed91c)
Signed-off-by: Ismael Juma <ismael@juma.me.uk>