KAFKA-3818: Change Mirror Maker default assignment strategy to round robin
It might make more sense to use round robin assignment by default for Mirror Maker since it gives a better balance between the instances,
in particular when the number of Mirror Maker instances exceeds the typical number of partitions per topic.
There doesn't seem to be any need to keep range assignment since copartitioning is not an issue.