Add the capability to maintain the old task assignment when the node topology is not changing
During cluster restart or rolling restart, the node topology is not changing.  But the current TaskAssignor is not able to maintain a deterministic mapping, so about 20% partitions will be reassigned which would cause state repopulation.
When the node topology is not changed, we really just want to keep the old task assignment.

Add the code to check whether the node topology is changing or not:
- when the prevAssignedTasks from the old clientStates is the same as the new task list
- when there is no new node joining (its prevAssignTasks would be either empty or conflict with some other nodes)
- when there is no node dropping (the total of prevAssignedTasks from other nodes would not be equal to the new task list)

When the topology is not changing, we would just use the old mapping.