KAFKA-3594; After calling MemoryRecords.close() method, hasRoomFor() method should return false
This exception is occurring when producer is trying to append a record to a Re-enqueued record batch in the accumulator. We should not allow to add a record to Re-enqueued record batch. This is due a bug in MemoryRecords.java/hasRoomFor() method. After calling MemoryRecords.close() method, hasRoomFor() method should return false.

This is a backport to the 0.9.0 branch.

Author: Manikumar reddy O <manikumar.reddy@gmail.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>, Ismael Juma <ismael@juma.me.uk>

Closes #1547 from ijuma/kafka-3594-for-0.9.0