KAFKA-3890: Streams use same task assignment on cluster rolling restart
Current task assignment in TaskAssignor is not deterministic.

During cluster restart or rolling restart, we have the same set of participating worker nodes.  But the current TaskAssignor is not able to maintain a deterministic mapping, so about 20% partitions will be reassigned which would cause state repopulation.
When the topology of work nodes (# of worker nodes, the TaskIds they are carrying with) is not changed, we really just want to keep the old task assignment.

Add the code to check whether the node topology is changing or not:
- when the prevAssignedTasks from the old clientStates is the same as the new task list
- when there is no new node joining (its prevAssignTasks would be either empty or conflict with some other nodes)
- when there is no node dropping out (the total of prevAssignedTasks from other nodes would not be equal to the new task list)

When the topology is not changing, we would just use the old mapping.

I also added the code to check whether the previous assignment is balanced (whether each node's task list is within [1/2 average -- 2 * average]), if it's not balanced, we will still start the a new task assignment.

Author: Henry Cai <hcai@pinterest.com>

Reviewers: Guozhang Wang <wangguoz@gmail.com>

Closes #1543 from HenryCaiHaiying/upstream

(cherry picked from commit a34f78dcad1d7844a21d0afbf0f6eef183847d0d)
Signed-off-by: Guozhang Wang <wangguoz@gmail.com>