KAFKA-3802; log mtimes reset on broker restart / shutdown
There seems to be a bug in the JDK that on some versions the mtime of
the file is modified on FileChannel.truncate() even if the javadoc states
`If the given size is greater than or equal to the file's current size then
 the file is not modified.`.

This causes problems with log retention, as all the files then look like
they contain recent data to Kafka. Therefore this is only done if the channel size is different to the target size.

Author: Moritz Siuts <m.siuts@emetriq.com>

Reviewers: Jun Rao <junrao@gmail.com>, Ismael Juma <ismael@juma.me.uk>

Closes #1497 from msiuts/KAFKA-3802-log_mtimes_reset_on_broker_shutdown